cleaned py scripts, tested and refined string functions
